requirements.txt
python-telegram-bot==20.7
telethon==1.28.5
requests==2.31.0
aiohttp==3.8.5
flask==2.3.3
